re-worked the routing and implemented waiting page
